Code refactoring - create base class ImgUploaderTool for different storages
