+++ src/tools/storage/imgur/imguruploadertool.h
@@ -17,17 +17,14 @@
 
 #pragma once
 
-#include "src/tools/abstractactiontool.h"
+#include "src/tools/storage/imguploadertool.h"
 
-class ImgurUploaderTool : public AbstractActionTool
+class ImgurUploaderTool : public ImgUploaderTool
 {
     Q_OBJECT
 public:
     explicit ImgurUploaderTool(QObject* parent = nullptr);
 
-    bool closeOnButtonPressed() const;
-
-    QIcon icon(const QColor& background, bool inEditor) const override;
     QString name() const override;
     static QString nameID();
     QString description() const override;
@@ -35,11 +32,4 @@ public:
     QWidget* widget() override;
 
     CaptureTool* copy(QObject* parent = nullptr) override;
-    void setCapture(const QPixmap&);
-
-public slots:
-    void pressed(const CaptureContext& context) override;
-
-private:
-    QPixmap capture;
 };
